+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Features
 - **Breaking Change**: upgrade `graphql` to `^14.0.2`, `graphql-subscriptions` to `^1.0.0` and `@types/graphql` to `^14.0.2`
+- update all other dependencies
 
 ## v0.14.0
 ### Features