+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog and release notes
 
-## Unreleased
+## v0.15.0
 ### Features
 - **Breaking Change**: upgrade `graphql` to `^14.0.2`, `graphql-subscriptions` to `^1.0.0` and `@types/graphql` to `^14.0.2`
 - update all other dependencies